epic_events/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.2
django-filter==21.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
phonenumberslite==8.12.42
psycopg2==2.9.3
pytz==2021.3
sentry-sdk==1.5.5
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8